# worker/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_transcriptions_v1_audio_transcriptions_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Audio or video file",
            "title": "File",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "response_format": {
            "default": "json",
            "title": "Response Format",
            "type": "string"
          },
          "temperature": {
            "default": 0.0,
            "description": "Greedy decoding only: 0 is the sole accepted value",
            "title": "Temperature",
            "type": "number"
          },
          "timestamp_granularities[]": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp Granularities[]"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_transcriptions_v1_audio_transcriptions_post",
        "type": "object"
      },
...
      "EmbeddingsRequest": {
        "properties": {
          "encoding_format": {
            "const": "float",
            "default": "float",
            "title": "Encoding Format",
            "type": "string"
          },
          "input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Input"
          },
          "input_type": {
            "default": "document",
            "enum": [
              "document",
              "query"
            ],
            "title": "Input Type",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "required": [
          "input"
        ],
        "title": "EmbeddingsRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The envelope every handled failure uses.\n\nDocumented because it is the shape a client branches on, and because it is\n*not* FastAPI's `{\"detail\": ...}` \u2014 a caller that reads only `detail` gets\nnothing useful out of a 503.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VerboseTranscriptionOut": {
        "properties": {
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "title": "Segments",
            "type": "array"
          },
          "task": {
            "const": "transcribe",
            "default": "transcribe",
            "title": "Task",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "VerboseTranscriptionOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stateless inference worker for vidtheque: speech-to-text with word timestamps,\ntext embeddings, frame embeddings, and on-screen-text OCR. One lifecycle manager\nowns the GPU \u2014 requests queue rather than compete. The transcription and text\nembedding endpoints follow OpenAI's shapes so a GPU-less deployment can swap in a\nhosted provider. Text and frame embeddings are separate vector spaces with\nseparate models and separate endpoints; they are never interchangeable. A\nnatural-language query reaches the frame space through the frame model's own\ntext tower, on /v1/embeddings/frame-query \u2014 never through /v1/embeddings.\n",
    "title": "vidtheque-worker",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audio/transcriptions": {
      "post": {
        "operationId": "transcriptions_v1_audio_transcriptions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcriptions_v1_audio_transcriptions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerboseTranscriptionOut"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Shape follows `response_format`: `verbose_json` is the schema below (segments, and word timestamps when word granularity is asked for), `json` is `{\"text\": ...}`, `text` is the transcript as `text/plain`."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The payload is the problem \u2014 an undecodable image, an unsupported parameter. Re-sending it will fail the same way, so this is not a retry: `error.type` names which."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Batch larger than this endpoint's cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The backend failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not now: the model is loading, the card is full, the lease hook failed, a device error unloaded the slot, or the worker is shutting down. Carries `Retry-After` in seconds."
          }
        },
        "summary": "Transcribe audio (OpenAI-compatible)"
      }
    },
    "/v1/embeddings": {
      "post": {
        "operationId": "embeddings_v1_embeddings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The payload is the problem \u2014 an undecodable image, an unsupported parameter. Re-sending it will fail the same way, so this is not a retry: `error.type` names which."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Batch larger than this endpoint's cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The backend failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not now: the model is loading, the card is full, the lease hook failed, a device error unloaded the slot, or the worker is shutting down. Carries `Retry-After` in seconds."
          }
        },
        "summary": "Embed text (OpenAI-compatible)"
      }
    },
    "/v1/embeddings/frame-query": {
      "post": {
        "description": "Vectors are L2-normalised and comparable to `/v1/embeddings/image`\noutput \u2014 and to nothing `/v1/embeddings` returns.",
        "operationId": "frame_query_embeddings_v1_embeddings_frame_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrameQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The payload is the problem \u2014 an undecodable image, an unsupported parameter. Re-sending it will fail the same way, so this is not a retry: `error.type` names which."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Batch larger than this endpoint's cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The backend failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not now: the model is loading, the card is full, the lease hook failed, a device error unloaded the slot, or the worker is shutting down. Carries `Retry-After` in seconds."
          }
        },
        "summary": "Embed a text query into the frame vector space"
      }
    },
    "/v1/embeddings/image": {
      "post": {
        "description": "Vectors come back in upload order, L2-normalised, one per image.",
        "operationId": "image_embeddings_v1_embeddings_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_image_embeddings_v1_embeddings_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The payload is the problem \u2014 an undecodable image, an unsupported parameter. Re-sending it will fail the same way, so this is not a retry: `error.type` names which."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Batch larger than this endpoint's cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The backend failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not now: the model is loading, the card is full, the lease hook failed, a device error unloaded the slot, or the worker is shutting down. Carries `Retry-After` in seconds."
          }
        },
        "summary": "Embed images into the frame vector space"
      }
    },
    "/v1/ocr": {
      "post": {
        "description": "Exactly one `data` entry per uploaded file, in upload order.\n\nAn image the decoder refuses gets an entry with `error` set and no items,\nrather than failing the batch: one unreadable keyframe out of sixty-four\nshould not cost the other sixty-three their OCR.",
        "operationId": "ocr_v1_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ocr_v1_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The payload is the problem \u2014 an undecodable image, an unsupported parameter. Re-sending it will fail the same way, so this is not a retry: `error.type` names which."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Batch larger than this endpoint's cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The backend failed unexpectedly."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not now: the model is loading, the card is full, the lease hook failed, a device error unloaded the slot, or the worker is shutting down. Carries `Retry-After` in seconds."
          }
        },
        "summary": "OCR one or more images"
      }
    }
  }
}